--- components/editor/DnDPrimitiveTypes.tsx
@@ -1,9 +1,0 @@
-const DnDPrimitiveTypes = {
-  PRIMITIVE_BUTTON: 'PRIMITIVE_BUTTON',
-  PRIMITIVE_INPUT: 'PRIMITIVE_INPUT',
-  PRIMITIVE_MOONMAIL_CONNECTOR: 'PRIMITIVE_MOONMAIL_CONNECTOR',
-  PRIMITIVE_FORM: 'PRIMITIVE_FORM',
-  PRIMITIVE_CONTAINER: 'PRIMITIVE_CONTAINER',
-}
-
-export default DnDPrimitiveTypes
