+++ components/editor/DragItem.tsx
@@ -1,48 +1,33 @@
-import { FC, useContext } from 'react'
+import { FC } from 'react'
 import { useDrag } from 'react-dnd'
-import DnDPrimitiveTypes from './DnDPrimitiveTypes'
-import { GlobalContext } from 'context/GlobalContext'
-import PrimitiveContainer from './PrimitiveContainer'
-import PrimitiveInput from './PrimitiveInput'
-import PrimitiveButton from './PrimitiveButton'
-import PrimitiveForm from './PrimitiveForm'
-import PrimitiveMoonmailConnector from './PrimitiveMoonmailConnector'
 
-interface DragItemProps {
-  type: string
-}
+import { DnDPrimitiveTypes, DragItemProps } from './types'
+import PrimitiveButton from '@/components/editor/primitives/PrimitiveButton'
+import PrimitiveContainer from '@/components/editor/primitives/PrimitiveContainer'
+import PrimitiveForm from '@/components/editor/primitives/PrimitiveForm'
+import PrimitiveInput from '@/components/editor/primitives/PrimitiveInput'
+import PrimitiveMoonmailConnector from '@/components/editor/primitives/PrimitiveMoonmailConnector'
+
 const DragItem: FC<DragItemProps> = ({ type }) => {
-  const { dispatch } = useContext(GlobalContext)
-  // console.log(type)
   const [, drag] = useDrag(() => ({
-    type: type,
     item: { name: type },
-    // end: (item, monitor) => {
-    //   const dropResult = monitor.getDropResult()
-    //   if (item && dropResult) {
-    //     console.log('dropResult', dropResult)
-    //     dispatch({ type: 'DROP_ITEM', payload: item.name })
-    //   }
-    // },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
+    type: type,
   }))
+
   return (
     <div ref={drag}>
       {(() => {
         switch (type) {
+          case DnDPrimitiveTypes.PRIMITIVE_BUTTON:
+            return <PrimitiveButton />
           case DnDPrimitiveTypes.PRIMITIVE_CONTAINER:
             return <PrimitiveContainer />
-          case DnDPrimitiveTypes.PRIMITIVE_MOONMAIL_CONNECTOR:
-            return <PrimitiveMoonmailConnector />
           case DnDPrimitiveTypes.PRIMITIVE_FORM:
             return <PrimitiveForm />
           case DnDPrimitiveTypes.PRIMITIVE_INPUT:
             return <PrimitiveInput />
-          case DnDPrimitiveTypes.PRIMITIVE_BUTTON:
-            return <PrimitiveButton />
-
+          case DnDPrimitiveTypes.PRIMITIVE_MOONMAIL_CONNECTOR:
+            return <PrimitiveMoonmailConnector />
           default:
             return null
         }
