+++ components/editor/DragItemWithDropZone.tsx
@@ -1,65 +1,50 @@
 import { FC, useContext } from 'react'
 import { useDrag, useDrop } from 'react-dnd'
-import DnDPrimitiveTypes from './DnDPrimitiveTypes'
-import { GlobalContext } from 'context/GlobalContext'
-import PrimitiveContainer from './PrimitiveContainer'
-import PrimitiveInput from './PrimitiveInput'
-import PrimitiveButton from './PrimitiveButton'
-import PrimitiveForm from './PrimitiveForm'
-import PrimitiveMoonmailConnector from './PrimitiveMoonmailConnector'
-import DropZone from '@/components/editor/DropZone'
 
-interface DragItemWithDropZoneProps {
-  type: string
-}
+import {
+  DnDPrimitiveTypes,
+  DragItemWithDropZoneProps,
+} from '@/components/editor/types'
+import { GlobalContext } from '@/context/GlobalContext'
+import PrimitiveContainer from '@/components/editor/primitives/PrimitiveContainer'
+import PrimitiveForm from '@/components/editor/primitives/PrimitiveForm'
+import PrimitiveMoonmailConnector from '@/components/editor/primitives/PrimitiveMoonmailConnector'
+
 const DragItemWithDropZone: FC<DragItemWithDropZoneProps> = ({ type }) => {
   const { dispatch } = useContext(GlobalContext)
   const [, drag] = useDrag(() => ({
-    type: type,
     item: { name: type },
-    // end: (item, monitor) => {
-    //   const dropResult = monitor.getDropResult()
-    //   if (item && dropResult) {
-    //     console.log('dropResult', dropResult)
-    //     dispatch({ type: 'DROP_ITEM', payload: item.name })
-    //   }
-    // },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
+    type: type,
   }))
-
-  const [{ handlerId }, drop] = useDrop({
+  const [, drop] = useDrop({
     accept: [DnDPrimitiveTypes.PRIMITIVE_BUTTON],
-    collect(monitor) {
-      return {
-        handlerId: monitor.getHandlerId(),
-      }
-    },
-    drop: (item, monitor) => {
+    drop: (item: any) => {
       dispatch({
-        type: 'DROP_ITEM_IN_ITEM',
         payload: {
-          id: type.name
-            .split('_')
-            .map(
-              (element) => element.charAt(0) + element.slice(1).toLowerCase()
-            )
-            .join(''),
           block: {
+            children: [],
+            connections: [],
             id: item.name,
+            props: {},
             type: item.name
               .split('_')
               .map(
-                (element) => element.charAt(0) + element.slice(1).toLowerCase()
+                (element: any) =>
+                  element.charAt(0) + element.slice(1).toLowerCase()
               )
               .join(''),
-            props: {},
-            connections: [],
-            children: [],
           },
+          id: type
+            .split('_')
+            .map(
+              (element: any) =>
+                element.charAt(0) + element.slice(1).toLowerCase()
+            )
+            .join(''),
         },
+        type: 'DROP_ITEM_IN_ITEM',
       })
+
       return { name: type }
     },
   })
@@ -70,10 +55,10 @@ const DragItemWithDropZone: FC<DragItemWithDropZoneProps> = ({ type }) => {
           switch (type) {
             case DnDPrimitiveTypes.PRIMITIVE_CONTAINER:
               return <PrimitiveContainer />
-            case DnDPrimitiveTypes.PRIMITIVE_MOONMAIL_CONNECTOR:
-              return <PrimitiveMoonmailConnector />
             case DnDPrimitiveTypes.PRIMITIVE_FORM:
               return <PrimitiveForm />
+            case DnDPrimitiveTypes.PRIMITIVE_MOONMAIL_CONNECTOR:
+              return <PrimitiveMoonmailConnector />
             default:
               return null
           }
