+++ components/editor/primitives/PrimitiveContainer.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-const PrimitiveContainer: FC = ({ children }) => {
+const PrimitiveContainer: FC = () => {
   return (
     <div className="overflow-hidden rounded-lg bg-lime-300 bg-white shadow">
       <div className="px-4 py-5 sm:p-6"></div>
